[WIP] Move density and full kplane fields in single file
